+++ Userland/sleep.cpp
@@ -6,7 +6,7 @@
 static unsigned parse_uint(const String& str, bool& ok)
 {
     unsigned value = 0;
-    for (size_t i = 0; i < str.length(); ++i) {
+    for (int i = 0; i < str.length(); ++i) {
         if (str[i] < '0' || str[i] > '9') {
             ok = false;
             return 0;
